Flytt noen funksjoner som hører til om barna i egen fil i tilhørende mappe
